Hygiene - move non-widget stuff to site
