+++ rmf_site_editor/src/widgets/inspector/inspect_model_description/inspect_mobility.rs
@@ -17,16 +17,13 @@
 
 use super::{
     get_selected_description_entity,
-    inspect_robot_properties::{
-        serialize_and_change_robot_property_kind, show_robot_property_widget,
-        RobotPropertyWidgetRegistry,
-    },
-    ModelPropertyQuery,
+    inspect_robot_properties::{show_robot_property_widget, RobotPropertyWidgetRegistry},
 };
 use crate::{
     site::{
-        Change, DifferentialDrive, Group, Mobility, ModelMarker, ModelProperty, RecallMobility,
-        Robot, RobotProperty,
+        robot_properties::serialize_and_change_robot_property_kind, Change, DifferentialDrive,
+        Group, Mobility, ModelMarker, ModelProperty, ModelPropertyQuery, RecallMobility, Robot,
+        RobotProperty,
     },
     widgets::{prelude::*, Inspect},
 };
